Fix more race conditions which cause Terminal Server SSH ERROR 'Bad argument, parameter 1' (-1) ... from output_thread

+++ sbbs3/js_global.c
@@ -1,6 +1,6 @@
 /* Synchronet JavaScript "global" object properties/methods for all servers */
 
-/* $Id: js_global.c,v 1.373 2018/03/10 06:12:50 deuce Exp $ */
+/* $Id: js_global.c,v 1.374 2018/04/06 02:47:06 rswindell Exp $ */
 // vi: tabstop=4
 
 /****************************************************************************
@@ -672,7 +672,7 @@ js_load(JSContext *cx, uintN argc, jsval *arglist)
  * It does assume the args are always last though (which seems reasonable
  * since it's variable length)
  */
-#define JS_ARGS_OFFSET	((unsigned)(JS_ARGV(0, (jsval *)NULL))/sizeof(jsval *))
+#define JS_ARGS_OFFSET	((unsigned long)(JS_ARGV(0, (jsval *)NULL))/sizeof(jsval *))
 
 static JSBool
 js_require(JSContext *cx, uintN argc, jsval *arglist)
